aven: Improve SVG output compatibility.  Tested with Mozilla Firefox 1.5, with Adobe's SVG browser plugin, Gimp, Gqview, and Opera 8.5.
aven: Put a 5mm border around exported SVG files to allow for station markers
and non-zero width lines.

aven: Fix crash when exporting as SVG or Sketch if labels or surface data
was turned on.

aven: Actually close the file we're exporting which fixes problems with it
not always being fully written.


git-svn-id: file:///home/survex-svn/survex/branches/survex-1_1@3191 4b37db11-9a0c-4f06-9ece-9ab7cdaee568