Fix jump to error for filenames containing colons
If an error only had a line number (but no column), and the filename
contained a colon (most do on MS Windows, as a full path usually has
a drive letter), we'd split at that colon, truncating the filename,
getting a bogus line number, and using the line number as the column
number.

Now fixed so we always include a second colon even if there's no
column number, so we can unambiguously split it up again.

Bug noted by Jenny Black.