Ensure that pcs->begin_survey is initialised
read_prefix may longjmp and so avoid the subsequent assignment.

This is a targeted fix for test failures visible in some environments
and under valgrind.  It may fix https://bugs.debian.org/821195.  A more
complete fix would be to make sure that structures are always
zero-filled when they are allocated, but that probably requires work
throughout the codebase.