Fix highlight of non-ASCII diagnostic type
When the translation of "error" or "warning" contained non-ASCII
characters, the code to highlight it in red or orange failed to
work.  The fix for this only works with a Unicode build of wxWidgets,
but as of wxWidgets 3.0, all builds are Unicode, so this shouldn't
be much of a problem as wxWidgets 2.x is close to obsolete now.

Reported by Mateusz Golicz.