Ignore view for export formats which work in 3D
When the rotation and tilt controls are hidden in the export dialog
we were still using their values to transform the data, so if you
set them with for one export format which support them, then
switched to an export format which doesn't, you'd get bogus
coordinates in the exported file.

Bug probably introduced in 1.2.27 by fixes for exports of rotated plans
and tilted elevations.