smoke.tst: Fix not to hang on macOS
Pass --help to aven when we run it to check that an unknown language
doesn't cause a segfault, and check for exit code 0.  Previously we
ran it with no options and with DISPLAY unset - under X that caused
it to exit with code 255, but on other platforms it would start
aven and then wait until a human closed it.