survexport: Replace --no-xxx options with --xxx
If no options specify what to export are given, use a per-format
default.  If the user specifies --defaults explicitly then the defaults
are always added so the user can easily build on the defaults.

Report any specified options which aren't supported for the format
being exported to.

Don't waste effort calculating bounds on the data if this information
isn't needed for the format being exported to (GPX, KML, POS).