Support "clamp to ground" for KML export
The default altitude mode for KML is "clampToGround", which renders
data on the surface of the terrain.  This is useful with KML viewers
which render the terrain as opaque so underground data isn't visible.
Rendering cave passages on the surface isn't great, but is better
than not being able to see them at all.

This option may also be helpful if you want to see where to look on the
surface for new entrances.