Improve handling of Compass MAK files
Handle fixed point coordinates in feet - previously the units were
ignored and the coordinates assumed to be in metres.

Previously the first byte in a MAK file was being ignored.  Typically
MAK files start with a comment, and since cavern currently ignores lines
that start with characters it doesn't understand the meaning of this
would often go unnoticed.