Deprecate MILS as angular units
Survex has long support MILS as an alias for GRADS.  However, this seems
to be a bogus definition of a "mil" which is unique to Survex (except
that Therion has since copied it) - there are several different
definitions of a "mil" but they vary from 6000 to 6400 in a full circle,
not 400.

Reported by Andy Edwards.