Avoid undefined signed shifts
Shifted a signed integer is undefined behaviour in some cases, so
make the type unsigned where this matters.

Survex itself is only affected on big-endian platforms, but this also
affects the img library on all platforms when used in standalone mode
as it is in other programs.

Fixes #119, reported by Matěj Plch.