Suppress lsan leak checking on exit
If the tree is built with LeakSanitiser then some tests fail due to
us not releasing all memory explicitly on exit (doing so would mean
extra work for no reason as the OS reclaims all memory when the
process exits).  If LeakSanitiser isn't in use then setting LSAN_OPTIONS
should be harmless.