aven: Fix handle of partly read UTF-8 sequences
When reading cavern output for the log window, we need special handling
for the case when a chunk of output ends mid-way through a UTF-8
sequence.  Previously we lost the first byte of the sequence in this
case (and would then show it as an invalid character), but now it is
handled correctly.

In practice, most of the cavern log output is ASCII so it's quite
possible nobody's ever actually hit this - I spotted it from looking
at the code.