Relax PROJ requirement to >= 6.2.0
PROJ 7.2.0 upgraded to EPSG 10.x which added the concept of a datum
ensemble, and this version of PROJ also added an API to deal with these.

If we're using PROJ < 7.2.0 then its EPSG database won't have datum
ensembles, so we don't need any code to handle them.

So we can relax the previous requirement of PROJ >= 7.2.0 and now the
requirement is determined by needing proj_create_crs_to_crs_from_pj()
which was added in PROJ 6.2.0.