Improve anon bound reporting
If the bound including anonymous stations is at an anonymous station but
the bound only considering named stations is the same, use the named
station for the anonymous bound too.

This helps to avoid reporting seemingly redundant bounds (though if the
anonymous bound is a fraction of a cm beyond the named bound then it'll
we'll still report what looks like a redundant bound).