[ci] Fix linux-latest build
It seems the Docbook DTDs aren't installed for some reason, so
ensure they are as that seems to solve the issue.

Also run the release upgrade in the home directory so any generated
files (such as "log") don't end up in the source tree and cause the
check for unknown generated files to fail.