Make "*fix" with no coordinates an info
This used to give:

warning: FIX command with no coordinates - fixing at (0,0,0)

However it's not really reporting a problem and the ability to omit the
coordinates is a deliberate feature.  It is useful for the user to know
where the "*fix" without coordinates is if they want to change the
survey to be in real coordinates, so an "info" diagnostic is a good fit.

Also if "*fix" is used twice with no coordinates we no longer omit this
diagnostic a second time right before the error:

error: Already had FIX command with no coordinates for station "x"