Accept weird .3d files from therion
Therion can create .3d files which have empty components in the
station name.  That's not really valid by Survex's definition of
station names, but it's not very helpful to reject such files, so
just disable the checks that were rejecting this.

The result is empty entries in the survey tree, but I don't see
what we can really do that's better than that.

Reported by Vasily Vl. Suhachev.