Fix two grid convergence bugs
One is due to a bug in PROJ < 9.3.0 with projected coordinate systems
with northing/easting axis order (such as EPSG:3042) which results in
the grid convergence being wrong by 90°.  We now work around this
problem if building with an affected PROJ version.  Reported by
Patrick Warren on the mailing list.

The other is that we weren't checking the output coordinate system
was set before calculating the convergence, which resulted in a
grid convergence of 0° being used if *declination auto was used
before *cs out.  We now handle this case correctly, and the
convergence is now calculated lazily which happens when we read a
compass reading, and gives an error if the output coordinate system
hasn't been set by then.