[img] Improve time_t conversion to days-since-1900
C signed integer division rounds towards zero, so previously
timestamps before 1970 would get rounded to the end of the day instead
of the start, so aside from midnight would end up one day later.

Now we offset before division so rounding is consistently to the start
of the day.

This only affects cases where we need to convert from time_t to
days-since-1900, and only for timestamps before 1970.