img: Improve PLT station reading
Eliminate duplicate img_LABEL items, which used to happen for
stations where more than two shots meet within a single survey.

We now infer img_SFLAG_WALL for the to station of a leg which
is flagged as a splay, since the Compass PLT format specification
documents this flag as meaning:

    The shot is a "splay" shot, which is a shot from a station to the
    wall to define the passage shape.

We now infer img_SFLAG_EXPORTED for any station that appears in more
than one survey in the PLT file.

We now set flag img_SFLAG_FIXED for any station which is listed as
a fixed point in the PLT file (a more recent addition to the format).