img: Support reading UTM zone & datum from PLT
Currently only datum "WGS 1984" is supported - if a different
datum is specified then we ignore the UTM zone and datum information
as before.

If the file specifies more than one datum or more than one zone
then we ignore this information as we'd need to re-project some
of the coordinates, which would require making PROJ a dependency.