img: Interpret DAT P flag as surface survey
This flag is described as "Exclude this shot from plotting", but the use
suggested in the Compass docs is for surface data, and they "[do] not
support passage modeling".

Even if it's actually being used for a different purpose, Survex
programs don't show surface legs by default so the end effect is at
least to not plot as intended.