[ci] Simplify linux-latest job
The idea of this job was to upgrade from GHA's "ubuntu-latest" to the
latest Ubuntu development version, but that upgrade isn't actually
happening now so strip out the code which tries to do this.  Originally
it was especially useful to allow testing with PROJ 8, but we get 8.2.1
from the standard ubuntu-latest now, and 9.1.1 from debian-stable.