Avoid deprecated proj_errno_string() function
Use proj_context_errno_string() instead if we're using PROJ >= 8.
The old function is "potentially thread-unsafe", but we don't use
it with threads so that's not actually a problem.  This is just
about future-proofing against the old function potentially giving
deprecation warnings or being removed in a future PROJ version.