Make showing caret in error reporting more robust
If due to a bug elsewhere we calculated a negative column offset, then
we'd print billions of spaces when trying to position the caret.
Noticed while developing on a branch - as far as I know this has never
happened in a released version.

Now a negative column offset is treated like zero and we don't try to
show the caret should this happen.