Reimplement aven's cavern log window
The old version turned the log into HTML and displayed it using
wxHtmlWindow, but that adds a lot of overhead and is especially
inefficient if there are many diagnostics - cavern could finish
almost instantly yet aven could take many seconds to process
the output.

The new version renders directly from the log data.  It should
have most of the features from before - the only missing feature
I'm currently aware of is that you can't now select and copy text
from the log window, which wxHtmlWindow provided for free.  If
people miss this feature, we could add it to the new implementation.