Eliminate Double typedef
It collides with an enum in the GDAL headers, and doesn't really
seem to be a useful distinction to have in our code (it's not used
consistently anyway).  Use GLdouble instead in the GL-specific
code.  In other code use double unless the type really needs to
exactly match GLdouble in which case use glaCoord (which is another
typedef we have for GLdouble).