Protect filenames which start with a dash
This applies when loading files from the cavern log window in aven.
We weren't doing this protecting on Microsoft Windows on the
assumption that it isn't needed there because the filenames will be
fully qualified, but that may not be the case if aven if run by hand
from the command line or from some other launcher.