Use a fixed rotation rate for pan to N,S,E,W
We were using the variable rate which auto-rotation does here, but
really these are just animations to help the user see the change
that's happening, so a fixed rate makes more sense (and is what tilting
to plan or elevation does).

The new fixed rotation rate for this is double what the default variable
rate was, which means the longest rotation (e.g. from S to N) takes
3 seconds rather than 6.