Convert manual from docbook to reStructured Text
The text-based format is much easier for making changes, and
for users the output from sphinx looks nicer and has extra
features (like built-in "Quick search" in the HTML output).

The plain text version of the manual is no longer generated.
Sphinx can produce one but the .rst sources actually seem
more readable for anyone wanting to read the manual in plain
text.

The extra targets to generate the manual in RTF and PostScript
have been dropped as I doubt they are still used (we stopped
providing them on the website years ago).  They could probably be
reinstated if they are still use.

As part of converting, the formatting of the manual has been
made more consistent.  Some content has also been reordered, and
other minor improvements made.