Fix missing report of unconnected survey piece
We aim to report at least one station in every unconnected piece
of survey, but if a piece had been simplified to a single station which
was anonymous (e.g. a disto splay shot) then we wouldn't report anything
for that piece.  Now we find the traverse that was attached to it and
report the next station along that traverse.