Handle an empty Walls station name
Walls allows a station with an explicit prefix to have an empty name,
e.g. `PEP:`.  The Walls documentation doesn't mention this, though it
also doesn't explicitly say the name can't be empty.  This quirk seems
unlikely to be intentionally used and Survex doesn't allow an empty
station name, so we issue a warning and use the name `empty name` (which
has a space in, so can't collide with a real Walls station name which
can't contain a space) - so `PEP:` in Walls becomes `PEP.empty name` in
Survex.

Reported by Erik C. Landgraf.