Give an error for `*fix a reference`
The "reference" token provides a way to have a list of known fixed
points which are not expected to all be currently used, while omitting
the coordinates provides a way to specify a point to arbitrarily fix
rather than rely on cavern picking one (which has the downsides of
the choice potentially changing when more survey data is added, and
of triggering an "info" message).  Doing both together doesn't really
make sense.