Only warn about hanging surveys
This has been an error previously, but at least for Walls data it seems
having hanging surveys is the norm, and Walls itself only warns about
them.

It seems we might as well support this for native Survex data too as
it allows viewing the connected parts of a survey with missing
connections without having to comment out the unconnected parts (and
then remember to fully uncomment once connections are surveyed).

If there are hanging surveys we may need to adjust the component
count, but don't currently so these statistics may be wrong so
for now we just suppress reporting them if there are hanging surveys.