Invent shorter names for anon stations in PLT export
The PLT file format documentation says station names can be up to
12 characters, but the ones we generated were longer than this
(e.g. 22 characters in pltexport.plt) which is liable to trip up
consumers of this format that only allow for the specified length.

We now generate names using a sequential counter with a two character
prefix so they should git in the 12 character limit for any cave
survey.