Handle "to" station starting `*` or `<` like Walls does
We were issuing an error about unterminated LRUD, but Walls seems
to check for a closing delimiter and if not found parse as a data
leg.  This leads to it (usually quietly) doing the wrong thing
when the closing LRUD delimiter is missing so we issue a warning
but provide an easy way to suppress that warning.