Remove Skencil export support
The last release of Skencil was in 2005.  There was an attempt to
revive the project in 2010, but that didn't lead to another release
and seems to have petered out.  No current Linux distro (or other
package system) seems to have packages for it.  The current git
version still appears to require Python 2.