Parse token as alphanumerics in most cases
Survex unintentionally allows omitting a token to be immediately
followed by a number, with no blanks in between.

We don't want to break files that rely on this (even if they likely only
do accidentally) but there are many cases where a number can't follow a
token and by parsing a contiguous sequence of alphanumerics we can
report better warnings or errors.

Partly addresses #135