Warn when a token is not followed by a blank
This is an unintentional tokenisation oddity which has been present
for a really long time.

We don't want to break files that rely on this (even if they likely only
do accidentally) so emit a warning (which can easily be eliminated by
added a space where indicated, which will work with old and new Survex
versions).

Fixes #135