Store short survey/station names inline
If a survey/station name is less than sizeof(char*)-1 we now store
it in the space where the pointer would be, and set bit
SFLAGS_IDENT_INLINE in sflags to indicate this.

This avoids an extra memory allocation for most stations on 32-bit
platforms (<= 3 bytes), and almost all stations and many surveys
on 64-bit platforms (<= 7 bytes), as well as improving data locality.

On a large dataset this reduced cavern's memory usage by ~4.5%
(3.5MB).  The speed-up was only measurable using a tool like
cachegrind that counts CPU cycles (~0.05% reduction).