Build without GDAL on Microsoft Windows
The build with GDAL currently doesn't work - aven fails during DLL
initialisation.  The problem seems to be in the packaged dependencies
but it's really unclear exactly where and the debugging tools don't
seem to help much.

We really need a working build so turning off GDAL for now seems the
least terrible option.  Hopefully the problem will get resolved for
us and we can turn this back on in the future.