Overhaul network articulation code
This fixes a bug with counting how many connected components the
survey network had, which lead to overreporting that value in
some cases (and also overreporting the loop count by the same
amount since the loop count calculation uses the number of connected
components).

The articulations are also now used to split up the problem a
little more, so we will often now solve more but smaller matrices
which should be faster.