Optimise validity check in remove_stn_from_list()
We optionally check if the station is actually on the list, which
we were doing by scanning the whole list for the station.  Instead
we can keep going back from the station to the list head and check
if that matches, which will scan only half the list on average.