aven: Improve "Colour by Survey"
The colours are based on hashing the survey name - we now calculate
the hash by iterating the survey name in reverse so that survey
names which differ only in the final character will tend to get
very different colours instead of quite similar colours.

Reported by Eric C. Landgraf and Patrick Warren.