Remove little endian fread/fwrite "optimisation"
Benchmarking shows it's actually faster to read/write 16 and 32 bit
integers byte-by-byte than with fread()/fwrite(), even if we don't
have getc_unlocked() and putc_unlocked().

With unlocked stdio on Linux, the time to read a large .3d file and
just count the stations and legs is more than halved by this change.