Use malloc()/realloc()/free() directly in img.c
This avoids needing osalloc.h in IMG_HOSTED mode, and we were only
using the wrappers which call the standard functions directly.

In the past, one motivation for these wrappers was platforms where
we needed to use a different allocator to support larger allocations
but no allocations in img.c should be very large and there are
probably no longer any such platforms (at least none which are relevant
to Survex).

They also provided a way to add memory allocation debugging (e.g. to
allocate with redzones) but there are various modern memory debuggers
which don't need source changes.