Strip out remains of support for defaultlang.h
This used to provide a way to build Survex such that the messages
used for errors encountered while trying to load the message file
could be translated.  We can't use the message file to get these
messages so these messages are compiled into the programs.

The machinery to create defaultlang.h was mostly removed in 2011,
but the machinery to use it left in place.  It doesn't seem worth
continuing to keep it around though - we only need one of these
messages if the installed files are not in the correct locations
or are corrupted, the system is extremely short on memory, or
a language is requested by setting SURVEXLANG which isn't
supported.

This simplification means we can just make the messages plain
ASCII so they don't need to be transliterated to the current
character set, which reduces the startup overhead a little.