CavernLogWindow now honours system dark mode
CavernLogWindow was using the desktop default window background
colour with black text, which was hard to read with a dark desktop
theme.

Now with a dark desktop theme CavernLogWindow uses white text on black;
with a light desktop theme the background colour is now set to white
(which matches how this worked when the log window was based on
wxHtmlWindow.