aven: Tighten workaround for ignoring error info
In some situations cavern currently fails to identify articulating
legs and writes out error information for them.  This error info
is always zero, and aven has a workaround to ignore it.  Previously
this only checked for E==0, but now we check that H==0 and V==0 too.

In theory this will help avoid "colour by error" from showing very
low error traverses as "not in loop", although in practice I've not
found a real world case where this makes a difference yet.