Turn off aven+survexport tests on mingw
We can't seem to run these programs during the build now.  This
appears to be due to some change in the Github CI images or a
dependency, as rebuilding the last good commit now fails.  There
is no error message, but bash shows exit code 127 which means it
failed to run the command.