Improve parsing of MAK files
Explicitly skip `/` comments, which seem to last until the end of
the line.  Previously a commented out MAK file command would still
be executed, which was incorrect.

Unknown commands are now warned about.  This seems more helpful as
it will help identify any bugs in Survex's parsing of MAK files,
and any undocumented MAK file commands which we don't support.